refactor(bench): extract bench artifact persistence into core service
The bench observation command owned the full bench-artifact persistence
engine — resolving artifact file pointers (shared-state and preserved-
invocation paths), recording file/directory/URL artifacts to the
observation store, rewriting artifact links back onto the results, and
recording budget findings. That is filesystem mutation + run artifact
persistence in the command layer, tripping the thin_command_adapter
audit detector.

Move the engine into core::extension::bench::artifact_persistence,
operating on the core ActiveObservation instead of the command's
BenchObservation newtype. The command file becomes thin wrappers that
unwrap observation.0 and delegate, preserving the existing
module-internal API (record_bench_observation_artifacts, record_if_exists,
record_memory_timeline_artifacts, apply_recorded_bench_artifact_links)
for lifecycle.rs and tests.rs.

Command module drops from 428 to ~33 lines.

Clears the bench/observation/artifacts.rs thin_command_adapter_violation.
